Fixed SD_Init issue + Project Version
Some MegaCommands would not initialise the SD Card if power was
switched off abruptly.

Seems to be an issue with the SPI timing of the  OLED and SD Card.

Moving the OLED init code to much later in start up sequence
has resolved this issue.

--

Project Version is now different to Firmware Version