Workaround for sent locks being saved as the parameter's default value.
Parameter locks are achieved by sending CC messages to the MD to
modify the parameter values of the current kit.

When the sequencer is running and paramter locks are being received
by the MD, the current kit values are always changing.

If the current kit is saved, the default kit paramters will be
overwritten by the last MIDI CC received.

MCL Saves the current kit in order to get the most up to date state
of the MD. However at a cost, it may save the parameter lock values
it has transmitted.

In order to work around this we do a few things.
- We dont save the current kit when the sequencer is running (one
exception is when we're in write mode ).
- We cache kit parameter values when the sequencer starts or when the
  parameter lock param type changes or when CC is sent from the MD.
- We reset parameter lock params when the sequencer is stopped by
  sending the cached values back to the MD.

A word about write mode. Write mode requires that the MD state be
saved. Mostly to obtain the current master fx settings. We can
probably assume the machine types are static from one write to the
next. Paramter changes are detected by MCL in realtime and are stored.
FX cant' be detected without saving the kit.
So when writing, is it okay to save the kit.
Yes: assume the user is not using locks on the md and locks from mcl
     assume that when the sequencer is running the user is using write
     mode for performance. when the sequencer stops values will be
     reset accordingly. the user can resave if they wish manually.