temp_track is now a generic Track object.
temp_track is a data structure used to previously store Track objects.
It's quite large. For efficiency i've created a parent EmptyTrack object
which is of a fixed size. Specific track types such as MDTrack or
A4Track can be loaded in to this data structure to eliminate space
wastage.

Improvements to GridPage GUI have been made.