Move to uart_block flag, to prevent MIDI overlap.
MIDI messages should not be interpolated.

For example, when a sysex message is sent, no other
data or status bytes should be sent.

Interpolation may occur when an ISR is triggered
during a multi byte MIDI message.

We now check in UART classes for sysex and multibyte
transmission events and set the uart_block flag.