Chain Mode Improvements:
Chain mode was incorrectly saving and requesting the current kit in all
circumstances. Should only do this if auto_save is enabled and if seq
is stopped.

When auto-chaining, chain mode will detect if the next machine/sound is
the same as the last one. If this is the case, it will not resend the
machine. This is useful if you are tweaking a machine over multiple
itterations of the sequencer, and want the tweaks to stay live.

A chain event initiated from the chain/write page will force the
machine/sound to be re-transmited.