Change note buffer data structure to bitmask with btree search.
Note buffer array data structure is used to store notes
that the sequencer has sent to device. Therefor when the sequencer
is stopped, note_off events for these notes can be sent.

The original implentation was an array which involved loops within
the seq isr. The new method avoids isr loops by taking advantage of bit masks.

To scan for active notes we use a btree search method.