Deprecate MutePage !Routing is now used for muting.
I've given up on obtaining the MD Mute state.

The audio routing functionality is an alternative
to muting and works well.

Removing the MutePage simplifies the firmware
and consolidates audio manipulation to 2 pages,
(MixerPage and RoutePage)).

Tracks can be "audio-muted' from the MixerPage
by holding "Write" and selecting tracks from
the MixerPage.