Refactor WIP
1) Consolidate load_from_grid functions from Track objects.

To simplfiy grid loading, we would like to be able to load any
track type from any slot position in the grid.

The parent track object is GridTrack. Let's make the parent object
responsible for loading tracks.

2) Store sequencer data, length and speed variables in GridChain
objects.

Previously, the length and speed of each sequencer track was contained
within the SeqTrackData object of each track type.

This meant that initialising these values, required calling functions
from with in child track objects.

In addition, chain mode requires knowing the length and speed of
a track, so it makes better sense co-locate these variables with chain
data.