Fixes for RAMPage(s)
Because we are now using the assignMachineBulk command we must initialise the
page 1 parameters ourselves.

Another consequence of assignMachineBulk is that if a REC machine is
changed to a playback machine before the record is finished, and the
parameter for CUE2 is modulated, the recording will be tainted with
feedback. Thus, we have to mute the recording track and wait a full
track record length before initating playback.

It appears microtiming was never set correctly for the RAM trigger
tracks resulting in shifted playblack.